# backend/docs/swagger.yaml
basePath: /api
definitions:
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  auth.LoginResponse:
    properties:
      token:
        type: string
    type: object
  auth.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the API for the HereWeGo application.
  title: HereWeGo API
  version: 0.0.1
paths:
  /:
    get:
      consumes:
      - application/json
      description: Returns a hello world message
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Returns a hello world message
      tags:
      - System
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: Login Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
      summary: Login
      tags:
      - Auth
  /auth/logout:
    post:
      description: Logout
      responses:
        "200":
          description: OK
      summary: Logout
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register
      parameters:
      - description: Register Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      responses:
        "200":
          description: OK
      summary: Register
      tags:
      - Auth
  /health:
    get:
      description: Returns the health of the database server
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Returns the health of the database server
      tags:
      - System
swagger: "2.0"